:construction: Code reformatting and naming.
