+++ src/main/resources/i18n/messages.properties
@@ -970,6 +970,10 @@ project.settings.metadata-templates.table.create-date=Created Date
 # ========================================================================================== #
 ProjectMetadataTemplates.title=Metadata Templates
 ProjectMetadataTemplates.fields={0} fields
+ProjectMetadataTemplates.create=Create New Template
+TemplateDetails.name=Template Name
+TemplateDetails.description=Description
+TemplateDetails.createdDate=Created Date
 
 TemplatePage.menu.details=Template Details
 TemplatePage.menu.fields=Metadata Fields
