--- src/main/webapp/resources/js/components/metadata-template/details/index.js
@@ -1,5 +1,0 @@
-import React from "react";
-import { render } from "react-dom";
-import { TemplatePage } from "./TemplatePage";
-
-render(<TemplatePage />, document.querySelector("#root"));
