+++ src/main/webapp/resources/js/components/metadata-template/template/index.js
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import { Layout, Tabs } from "antd";
 import { grey1 } from "../../../styles/colors";
 import { MetadataTemplateProvider } from "../../../contexts/metadata-template-context";
-import { TemplateDetails } from "./TemplateDetails";
-import { TemplateFields } from "./TemplateFields";
+import TemplateDetails from "./details";
+import TemplateFields from "./fields";
 
 const { Content } = Layout;
 
-export function TemplatePage({ children, templateId }) {
+export default function MetadataTemplatePage({ children, templateId }) {
   const [activeKey, setActiveKey] = useState("details");
 
   return (
@@ -16,12 +16,13 @@ export function TemplatePage({ children, templateId }) {
         <Content
           style={{
             backgroundColor: grey1,
-            paddingLeft: 20,
           }}
         >
           <Tabs onChange={setActiveKey} activeKey={activeKey}>
             <Tabs.TabPane key="details" tab={i18n("TemplatePage.menu.details")}>
-              <TemplateDetails />
+              <div style={{ padding: 15 }}>
+                <TemplateDetails />
+              </div>
             </Tabs.TabPane>
             <Tabs.TabPane key="fields" tab={i18n("TemplatePage.menu.fields")}>
               <TemplateFields />
