+++ src/main/webapp/resources/js/pages/projects/metadata-templates/index.js
@@ -1,5 +1,61 @@
 import React from "react";
 import { render } from "react-dom";
-import { ListMetadataTemplates } from "../../../components/metadata-template/list";
+import { navigate, Router, useLocation } from "@reach/router";
+import { setBaseUrl } from "../../../utilities/url-utilities";
+import ListProjectTemplates from "../../../components/metadata-template/list";
+import Template from "../../../components/metadata-template/template";
+import { Button, Layout, PageHeader } from "antd";
+import { grey1 } from "../../../styles/colors";
 
-render(<ListMetadataTemplates />, document.querySelector("#templates-root"));
+const { Content } = Layout;
+
+/**
+ * React component for setting up page routing on the project metadata template
+ * page.
+ * @returns {JSX.Element}
+ * @constructor
+ */
+function ProjectMetadataTemplatesPage() {
+  return (
+    <Router>
+      <PageLayout path={setBaseUrl("/projects/:projectId/metadata-templates")}>
+        <ListProjectTemplates path="/" />
+        <Template path={"/:templateId"} />
+      </PageLayout>
+    </Router>
+  );
+}
+
+const PageLayout = ({ children, projectId }) => {
+  const location = useLocation();
+
+  /*
+  Only show the back button if currently viewing a template.
+  The url will end in a number if viewing a template
+   */
+  const onBack = location.pathname.match(/-templates$/)
+    ? null
+    : () => navigate(setBaseUrl(`/projects/${projectId}/metadata-templates`));
+
+  return (
+    <PageHeader
+      className="site-page-header"
+      title={i18n("ProjectMetadataTemplates.title")}
+      onBack={onBack}
+      extra={[
+        <Button key="create-btn" className="t-create-template-btn">
+          {i18n("ProjectMetadataTemplates.create")}
+        </Button>,
+      ]}
+    >
+      <Layout>
+        <Content style={{ backgroundColor: grey1 }}>{children}</Content>
+      </Layout>
+    </PageHeader>
+  );
+};
+
+render(
+  <ProjectMetadataTemplatesPage />,
+  document.querySelector("#templates-root")
+);
